chore(Topic Salad): add docs to Word Salad and changing location of file
