+++ static/js/TopicLandingPage/TopicSalad.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {useEffect, useState} from "react";
-import {WordSalad} from "./WordSalad";
+import {WordSalad} from "../WordSalad";
 import Sefaria from "../sefaria/sefaria";
 import {InterfaceText} from "../Misc";
